requirements.txt
streamlit
pandas
spacy
pytesseract
scikit-learn
pdfplumber
pillow
numpy
opencv-python-headless
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
